feat: implement ABAC and refactor access control

+++ backend/src/main/java/com/atlas/dto/MonthlyAllocationDTO.java
@@ -14,5 +14,5 @@ public class MonthlyAllocationDTO {
     private Long allocationId;
     private Integer year;
     private Integer month;
-    private Double percentage;
+    private Integer percentage;
 }
